Update patch set 11
Patch Set 11: I would prefer that you didn't submit this

In MulticoreVisualizer.java, the updateLoads() code
is still invoking the DSF layer directly, rather than
going through DSFDebugModel.

I realize you guys are familiar/comfortable with DSF and hence may not see this as an issue, but I do think the code is cleaner if we don't mix the "view" and "model" layers together. This is particularly important if we're taking about refactoring this stuff to support contributions later.

(See my previous comment, copied below for reference.)

- The DSFSession/DsfExecutor/Runnable/updateLoads()
code should be folded into the existing
DSFDebugModel/DSFSessionState mechanism.
One reason for the DSFDebugModel being there
is to encapsulate this kind of repetitive DSF
session/executor stuff, and keep it from creeping
into the view code, so the view code can be expressed
cleanly in terms of the data it's looking for.
(You're already doing this with getLoad()/getLoadDone(),
it just feels like there's a layer of DSF code above
this that wants to be similarly encapsulated.)

Patch-set: 11