Update patch set 3
Patch Set 3:

No, I don't think so. We're backporting these fixes to an internal repo and reverting and reapplying this last round was painful. We can address concerns with another change request. This feature continues to be under development and I want to make sure we're focused on getting the functionality right.

Patch-set: 3