Update patch set 1
Patch Set 1: (1 inline comment)

Is there a bug # associated with this? If so, please prefix the commit message with it, like so:
Bug ######: blah blah blah

Also, please don't mix whitespace and sort-order changes in with substantive changes. I'm a big fan of cleaning up whitespace and out of order stuff, but mixing it with other changes makes the diffs unreadable.

Please upload a new patch set without the whitespace changes.

Patch-set: 1