Bug 402177: Btree.insert returns wrong value
The javadoc for BTree.insert says "don't insert if the key was already
there, in which case we return the record that matched".

However, the implementation was returning the new record even when it
was not actually inserted.

This is a fix for the problem and a test case to demonstrate the issue.

Change-Id: If079f235871fcdfbd35f1cba3f64cc3e33edaaec