Update patch set 1
Patch Set 1:

Anna, while it may or may not be documented, the reason one review is used is that it is very hard for reviewers to track comments from the previous patch or patches and evaluate how they were handled when dealing across multiple gerrit reviews.  With one review, all comments and responses for all patches can be viewed in one place and thus the history of the patch can be inspected.  It may take several iterations to make everyone happy and sometimes it means going back to a previous iteration.

Patch-set: 1