Update patch set 1
Patch Set 1:

I'm sorry -- there was a misunderstanding.

There's certainly no rule against having multiple changesets per bug (see Bug 402071, one of mine, which has had 4 changesets so far). But, as Jeff said, it's much more helpful to reviewers to have different versions of the *same* change in the same gerrit entry, so they can be more easily compared. That was all I was asking for. 

I do apologize if my tone came off as too harsh or high-handed.

Here's an example of the neat comparisons that are possible when all the versions of a change are in the same gerrit entry: https://git.eclipse.org/r/#/c/10762/1..4/doc/org.eclipse.cdt.doc.user/reference/cdt_u_m_file.htm -- This is a comparison between the 1st and 4th versions -- note the "1..4" part of the URL.

Patch-set: 1