Update patch set 2
Patch Set 2:

Clearly I'm an idiot -- K&R is obviously a C/C++ style, when I looked in the Java style section I found "Eclipse built-in", which is what I used.

However, I still think this would be a good topic for the wiki (if it isn't already there).  If it is there, then perhaps it should be more prominent?

Patch-set: 2