Update patch set 1
Patch Set 1:

With this patch the persistence of expanded state will be based on the function name.  So if you have variables expanded, then you step into a function, your expanded state will be reset, but if you select the frame one level up, the old expansion states should be restored.

Patch-set: 1