Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

That's a fine way; in the future, putting it right at the beginning of the commit message (like "Bug XXXXX: one-line description of commit") is more typical, and enables people to look at the bug when they are viewing single-line commit messages.

Thanks for updating it.

Patch-set: 2