Bug 403382: Find References within QObject::connect
Qt marks some functions as signals or slots, e.g.,

    class T : public QObject
    {
    Q_SIGNAL void signal( int );
    Q_SLOT   void slot();
    };

These special functions are later linked by passing them to the
QObject::connect function, e.g.,

    T t1, t2;
    QObject::connect(
        &t1, SIGNAL( signal( int ) ),
        &t2, SLOT( slot1() ) );

Selecting either the signal or slot function declarations and then
choosing 'Find References' does not find the reference in the
QObject::connect function call.

This patch adds new functionality to the cdt.core indexer.  It allows
contributors to provide extra IASTNodes as children of existing nodes.

The Qt changes in this patch use the extension point to recognize
the QObject::connect function call and then create new children
for the SIGNAL and SLOT expansions.  The children are IASTName's
that reference the IBindings for T::signal(int) and T::slot().

My initial implementation tried to create children for the SIGNAL and
SLOT macro expansions.  However, I couldn't find an efficient way to get
from the expansion node back to the function call expression.  This
meant that I couldn't properly ignore signal functions from unrelated
types.  The submitted solution recognizes the function call expression,
so it is able to find the right function within the type of the
parameter that is being connect'ed.

Change-Id: Id1eb9198d1fc41e57bfd027a42e587dbd25378fb