[Bug 317353] Reduce space requirements of various prefs pages
The area covered by CDT prefs pages has been always quite huge. Often,
the scroll bar of the virtual composite has to be displayed although the
components could display all the required information readily in a
smaller box. This is an attempt to reduce some of the fixed space
requirements. The solution is not ideal, but the layout framework of SWT
(no notation of minimum size) and the way how sizes are computed in
prefs page by jface doesn't allow a more sophisticated solution at this
time.

Change-Id: I5f9aa36d0d37b93852a39d636719be801a2c1360
Reviewed-on: https://git.eclipse.org/r/11299
Reviewed-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>
IP-Clean: Sergey Prigogin <eclipse.sprigogin@gmail.com>
Tested-by: Sergey Prigogin <eclipse.sprigogin@gmail.com>