Update patch set 1
Patch Set 1: Do not submit

Whitespace/formatting errors ought to be fixed.  It's easy enough, just run it through the formatter.

The duplicate if/else clauses should be fixed (although this was a problem with the original code).

I would like one of the debug committers to please have a look.

Patch-set: 1