Update patch set 3
Patch Set 3: (12 inline comments)

Since some of my comments were never responded to, I'll repeat them on this version. They are only suggestions, but I would appreciate a response at least.

And lots more places where if or for statements are missing braces -- this is asking for errors down the line...

Patch-set: 3