Update patch set 3
Patch Set 3:

I ran this fix with two processes, one 64 bit and one 32 bit and I see the commands being sent properly and the correct answers received.  GDB then got confused about threads, but I think it is because my version is not compiled quite right for this scenario.  But it does work as we expect in this case.

Whenever adding a new process to the debug session, a new "show endian" is being sent, but we don't specify the process it should apply to.  I've put a comment about that in the code.

Patch-set: 3