Update patch set 4
Patch Set 4:

Marc, I have changed CLIShowEndian as you suggested. Please, let me know if this is what you wanted me to do.
I originally added John to the authors list, but it seems that only one author is allowed. So now John is the only author of the patch which may be a problem because the patch is not very small. Not sure what to do with this.

Patch-set: 4