Update patch set 1
Patch Set 1:

> The view expanded state is updated when the input 
> to the view changes, however if you're stepping 
> (or even run to a breakpoint fast), the input 
> stays the same even if you step into another 
> function. I.e.
> frame_before_step[0].equals(frame_after_step[0]) == true.

I don't quite get how the frame before the step can
be the same as the frame after the step, if we step
into a function, but I believe that is what I see
also.

This happens in the platform code or DSF?

> Unless we are willing to mess with the frame DMC's 
> equals or using some other hacks this workaround 
> won't work. I'm sorry I hadn't thought of it sooner.

If you feel this is too complicated, should we abandon this change?

Patch-set: 1