Bug 406652: PDOMTagIndex will not load tags from cold cache
The guard condition in PDOMTagIndex.getIdRecord was too aggressive. 
There was a check to prevent database reads when createIfNeeded was
false.  This is incorrect since reading from the database does not try
to create anything.  There is a guard later in the same method around
the code that does try to allocate a new record.

I've also added a test case to confirm the fix.

Change-Id: I03b5f94276756a74540144fa5231974ba429a68b