Update patch set 3
Patch Set 3:

Hi Bill,

Here is the latest patch.   I have made one small functional change: now the filter stays in place even if the debug session changes / the model is reconstructed.  

I decided to go this way because there is an annoying case (bug?) still where the model is reconstructed unexpectedly that would clear any filter put in place.  This happens when all threads are unselected in the multicore visualizer and the user re-selects a frame in the same session in the debug view.  What do you think? 

Regards,

Marc

Patch-set: 3