Update patch set 3
Patch Set 3:

Yes, a proper API would be good.  I would like to defer that to when we implement the "show exited processes in debug view", so that we get that API right.

The current patch was really just a quick way to get this to work for Kepler.

Patch-set: 3