Update patch set 7
Patch Set 7:

I modified my last patch so that the expression "*a=1" is marked as read occurrence for the variable a to avoid problems with refactoring. The tests were adapted accordingly.
I still think that "*a=1" should be marked as a write occurrence but I have to look for another solution. I will open a new bugzilla request if I am able to solve it properly (w/o breaking refactoring).

Patch-set: 7