Update patch set 2
Patch Set 2: Verified; Looks good to me, approved; IP review completed

I've updated Roland's so it could be committed right away.

I'm not thrilled with the fix, but I think it is the best we can so late in the cycle.  We'll improve on it for the next release.

Patch-set: 2
Label: SUBM=+1