Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve; IP review completed

(1 inline comment)

Doug, thanks for fixing this. I only have a tiny comment but not enough to hold this back so I put +1. I put 0 for verified because executables work (mac, linux) but if I try to create a shared library, I get
/usr/bin/ld: cannot find -link-as-library
but as you said, it's a stop gap. So I'll leave it up to you to either try to fix it part of this change or as a separate one.

Patch-set: 1