Update patch set 1
Patch Set 1: Code-Review-1

Is there a bug opened for this?
I'd like to know how to reproduce the problem.  We are going to have to change the patch to work for multi-process, but it is not clear what is happening with GDB, so I don't know what to do yet.

I'm hoping more info in a bugzilla will clarify.

Thanks

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1