Bug 400033 - gdb remote attach: automate finding local exe
Added in Launch Configuration Dialog, Attach to proceess, Debugger tab a Sysroot subtab.
The new subtatn is visible only when gdbserver is selected.
Added to read the provided sysroot value and to use it.
In case the selected process has an absolute path, its path is appended to sysroot path and the process is attached automatically.
In case the selected process has a relative path, the provided sysroot path is used for mapping remote paths to local paths.

Change-Id: Ib3dcb748ec10f5582fd085d2662418b75976fe46
Signed-off-by: Lidia Gutu <lidia.gutu@windriver.com>