Update patch set 1
Patch Set 1:

This solution breaks an existing test containing a scattered array. Before, the function-like macro calls got reformatted to the beginning of the next line, which does not look correct to me. If the macro in this test is replaced by a function call the result of the formatter looks similar to the new result. Therefore, I expect the new result to be better as well.

Patch-set: 1
Label: Verified=0