Update patch set 1
Patch Set 1: Code-Review-2

You are assuming that there is always 1 to 1 relation between language ID and its name. It is not guaranteed to be so. The case when different languages are defined with the same name needs to be considered. There is also the case when language is null which was covered on master after you branched. Generally, please rebase your patches against the latest master.

Patch-set: 1
Label: Code-Review=-2