Update patch set 2
Patch Set 2:

(3 comments)

>I think this could go in org.eclipse.cdt.releng, what do you think?

It's possible although the contents of this plugin have a different purpose, I think is best to keep target definitions in its own bundle which is consistent with other projects e.g. Mylyn and eGit.  at the same time we have to make them easy to be found.  Let me know your opinion on this.

> The bundle org.eclipse.test.performance is not resolved in org.eclipse.cdt.tests.dsf/META-INF

I have not found any update site that delivers o.e.test.performance, it seems we don't have a choice but to import this project into the work space.

>Running Eclipse 4.3 I have my baseline set to 4.3 / CDT 8.2 and my target platform to 3.8. I get several api errors lile: The super interfaces set has been reduced for type org.eclipse.cdt.launch.ui.CDebuggerTab. If I change to the 4.3 target, the api errors are gone. Any idea why that happens? Running 3.8.2 with target platform of 3.8 also has the same api errors.

I think this is related to the fact that today we can not specify API baselines within the target definitions (there is a bug open for that, and hopefully addressed in the near future), So we need to move our baselines manually as needed depending on the branch checkout on the work space.
We can take a closer look at this to find the explanation for it.

Patch-set: 2
Label: Verified=0