Bug 400033 - gdb remote attach: new sysroot subtab on launch configuration dialog.
Added a Sysroot sub tab in Launch Configuration Dialog-> Attach to
process-> Debugger tab. It is visible only when gdbserver is selected.
Added to read the provided sysroot value and to use it for automate
finding local exe.
- In case the selected process has an absolute path, its path is
appended to sysroot path and the process is attached automatically.
- In case the selected process has a relative path, the provided sysroot
path is used for mapping remote paths to local paths.

Change-Id: I67f5c9f8062bbccc2dcc95ad8ddf517570cab9ce
Signed-off-by: Lidia Gutu <lidia.gutu@windriver.com>