Bug 415789: NPE in ScannerConfigBuilder
An implementation of IEnvironmentVariable is returning null for its key
or value.  The javadoc on this interface doesn't mention null, which I
guess makes it a valid value.

This patch checks the result before trying to put it into an instance of
java.util.Properties.

Change-Id: Ic04ddd72dfb558ca403b549b64847c3437971407
Signed-off-by: Andrew Eidsness <andrewe@jfront.com>
Reviewed-on: https://git.eclipse.org/r/15820
Reviewed-by: Andrew Gvozdev <angvoz.dev@gmail.com>
IP-Clean: Andrew Gvozdev <angvoz.dev@gmail.com>
Tested-by: Andrew Gvozdev <angvoz.dev@gmail.com>