Added initial and limitted support for parsing doxygen comments.
Two new IAST types are introduced. IASTDoxygenComment is a specialization of IASTComment and
contains a list of IASTDoxgenTag instances. Tag name and values are stored as plain text and no
further structure is parsed. Currently, only Javadoc like doxygen comments can be recognized. That
is, the tag must start with an '@' and the comment itself must of the form /** */.

Signed-off-by: Sebastian Bauer <mail@sebastianbauer.info>