Update patch set 3
Patch Set 3:

It would definitely be be useful to have a class similar to DebugStringVariableSubstitutor as part of org.eclipse.core.variables plug-in. Since the platform is a slow-moving animal, I'm going to submit this change as is and then try to move this functionality to org.eclipse.core.variables.

Patch-set: 3
Label: Verified=0