Bug 416471: Support for extra ILanguageSettingsProviders in new projects
This change introduces three new ProcessRunners that can be used in the
New Project wizard's template.xml files.  These will be used by two new
Qt project wizards that I will introduce in a second patch.

The three new rules are:

1) "AddMakeTarget" which creates new Make Targets (in the Make Targets
view) for the new projects.

2) "SetEnvironmentVariable" which sets an environment variable in all of
the new project's build configurations.

3) "ExtraLanguageSettingsProvider" which modifies the new project's
build configurations to include a new ILanguageSettingsProvider.

The first two are straightforward, the third is a bit different.
Instead of creating a new Toolchain or Configuration it modifies the
Configurations that were created for the new project.  In this case the
only modification is to add the extra ILanguageSettingsProvider, but it
might be useful to extend this to other customizations as well.

Change-Id: I30710400e9b0dffcbe6e8965ce7ce2078c1c99ca
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>