Bug 418406: Qt4 and Qt5 specific HelloWorld project wizards
This replaces the QtQuick2 project wizards with a Qt5 wizard that
includes the old content as well as:

1) Make targets to build and clean the project
2) Sample interaction between C++ and QML (a trivial clock)
3) A Qt header path provider so that paths from the Qt installation are
properly resolved in the project.

Item 3 is particularly important for using the 'New Class' wizard with
QObject as the base class.  Unless the Qt paths are pre-populated the
QObject base class will not be found and the New Class wizard's Finish
button will be disabled.

The Qt headers are resolved by running `qmake -query QT_INSTALL_HEADERS`
and then creating IncludePath entries for all sub-folders.  This list of
include paths is stored in-memory which means they are computed once
(for each version of qmake) in each workbench session.

Change-Id: Ic82fdb147e6a69060f93e2e9aed2e919139a0ae9
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>