Update patch set 1
Patch Set 1: Verified-1

Actually, as Marc pointed out, gerrit does generate a Change-Id but not in the commit message, it's in the fields on the left, I missed that :)

As for the patch, I noticed a bad side effect. The first time the enhanced hover is shown in a new workspace, it is very tiny. When it is shown again it goes back to normal size. I'm not sure why that would happen, maybe the dialog settings are getting loaded with wrong values. I'll investigate.

Patch-set: 1
Label: Verified=-1