Update patch set 2
Patch Set 2:

With the previous patch, the first time computeSizeConstraints is called without the presence of a IDialogSettings, it calls computeSizeHint and ends up using the layout to compute the size, which shrinks the control. It didn't shrink before in this case because computeSizeConstraints returned the default (602x182) and that result affected computeSizeHint. What the patch was really meant to do is to use the size stored in the dialog settings as constraints when it exists, so in the new patch, I explicitly use the dialog settings instead of calling computeSizeHint.

Patch-set: 2