Update patch set 1
Patch Set 1:

Hi Sergey,

Sorry it took so long to get back to you, as we just had a holiday here. 

And about the testcase, I have found the CPartitionerTest, and I found that I can not simply add a test case, since if I want the correct partition returned, I have to call my partitioner scanner to do it. I'm not that familiar with the framework, so the question is, if I want to call my own partitioner, should I add my partitioner and scanner in the test package? Or just add a jar?

Patch-set: 1