Update patch set 4
Patch Set 4:

Patchset 4 is rebased to the offer frame specif values.
This trigger a change so the registers under the user group are new instances rather that re-using the same.
  A change is a register value now needs to produce an event to notify of registers changed to all groups, this is needed as some register values depend on other registers, and the dependent registers could span multiple register groups.

A new test case was created to validate that registers can be updated from any register group and this update is reflected on all groups.

Persistence is performed on the launch configuration and applies to all processes available. 

A discussion was raised to see if we can move the context parent of Register groups from Process (Contaner) to ICommandControl.
  user groups could be at container level, however no the General Group, as this is specific to the process compiled architecture e.g. 32 bits vs 64 bits, etc.
So it makes sense to keep the groups at the Container level.

There is a drawback on the current solution when users update register groups from multiple processes or multiple sessions, the last session close will override the group layout performed on the earlier sessions.
  However this scenario is not common and in fact present in CDI. It's proposed to raise a bug on this and follow the needs to enhance this particular scenario.

Patch-set: 4
Label: Verified=0