Bug 418406: Qt4 and Qt5 specific HelloWorld project wizards
This replaces the QtQuick2 project wizard with Qt4 and Qt5 wizards that
include the old content as well as:

1) Make targets to build and clean the project
2) Sample interaction between C++ and QML
3) A Qt header path provider so that paths from the Qt installation are
properly resolved in the project.

Item 3 is particularly important for using the 'New Class' wizard with
QObject as the base class.  Unless the Qt paths are pre-populated the
QObject base class will not be found and the New Class wizard's Finish
button will be disabled.

The Qt headers are resolved by running `qmake -query QT_INSTALL_HEADERS`
and then creating IncludePath entries for all sub-folders.  This list of
include paths is persisted with other shared language settings into a
file in the workspace metadata area.

The persisted data is reloaded when any of the following change:
- the modification time of the qmake binary
- the modification time of the reported QT_INSTALL_HEADERS folder

The persisted node is ignored when the target qmake binary no longer
exists.  The node is removed from the persisted form the next time that
the shared settings are persisted.

Change-Id: Ic82fdb147e6a69060f93e2e9aed2e919139a0ae9
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>
Reviewed-on: https://git.eclipse.org/r/16909
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>
Tested-by: Doug Schaefer <dschaefer@qnx.com>