Bug 418406: Move AbstractBuiltinSpecsDetector (and related) to cdt.core
AbstractBuiltinSpecsDetector, AbstractBuiltinSpecsDetector, and
AbstractLanguageSettingsOutputScanner are in the managedbuild plugin but
do not really depend on any functionality of that plugin.  This change
moves the classes into cdt.core so that they may be reused by clients
that don't otherwise need managedbuild.

I've updated all references to the new location.
I've replaced references to all classes.

Change-Id: Ib2d216878301197136be627a8ba11f79b1c20366
Also-by: Andrew Gvozdev <angvoz.dev@gmail.com>
Signed-off-by: Andrew Gvozdev <angvoz.dev@gmail.com>
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>