Bug 418406: Move AbstractBuiltinSpecsDetector (and related) to cdt.core
AbstractBuiltinSpecsDetector, AbstractBuiltinSpecsDetector, and
AbstractLanguageSettingsOutputScanner are in the managedbuild plugin but
do not really depend on any functionality of that plugin.  This change
moves the classes into cdt.core so that they may be reused by clients
that don't otherwise need managedbuild.

I've updated all references to the new location.

I've replaced references to all classes.  This change raises an API
exception.  I was able to replace two of the old classes with
implementations that inherit from the new classes.  I marked these two
classes with @deprecated and included a pointer to the new
implementation.  This was not possible for the third class because of an
internally declared enum; I've create a API filter for that case.

Change-Id: Ib2d216878301197136be627a8ba11f79b1c20366
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>