Bug 421041 - NPE in EvalBinding.getBinding
This happens when constructor EvalBinding(IBinding binding, IType type,
IBinding templateDefinition) is called with a null binding. A binding
can be null for example when TypeMarshalBuffer.unmarshalBinding returns
null. Because the binding is null when calling this constructor, the
assumption in EvalBinding.getBinding that fParameterOwner won't be null
isn't valid anymore.

Change-Id: Icebf875e059f2962cc2ddd91d3b79c51b88eddac
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>