Update patch set 4
Patch Set 4: Code-Review-2

(50 comments)

Thanks for this patch, the feature looks really good.

Sorry for the delay though.  This is a large change so I wasn't able to get through all of it but I got through most.

The first thing to fix is the handling of the decoupled actions.  Currently, they collide with each other.  I suggest using commands instead of actions.  But I put some detailed comments in-line.

Patch-set: 4
Label: Code-Review=-2