Update patch set 5
Patch Set 5:

For 8.2.2, we are not supposed to add new APIs, so I don't think we can commit as is.  However, I think we should fix it since it breaks completely the launch.  I suggest we just accept the error and let the launch continue.

Phil, would that be enough?

Patch-set: 5