Update patch set 1
Patch Set 1:

"Could we log a warning when the constructor in question is called with a null binding? I think that indicates a bug elsewhere in the code."

Sorry for the slow reply, I got side-tracked quite a bit. It looks like you're right. I looked into bug 418770 because it has small sample code and it has same NPE but different call stack. In this case, it was a bug with the enum specialization handling, see https://git.eclipse.org/r/#/c/19031/

But even with this other patch, the NPE still occurs in the case of Bug 421041 (using Boost source files). So it looks like the NPE is symptomatic of different bugs that are worth looking into and it could be logged. But I'd like to find the real issue with this case first.

Patch-set: 1
Label: Verified=0