Bug 422841: Add Q_CLASSINFO to the QtIndex
Qt allows string-based key/value pairs to be inserted into QObject class
definitions.  E.g.,

    class Q : public QObject
    {
    Q_OBJECT
    Q_CLASSINFO( "key1", "value1" )
    };

The class info is accessible in the meta-object system.  See:

    http://qt-project.org/doc/qt-4.8/qmetaclassinfo.html

For more information.

This patch adds the API to access these key/value pairs from the
QtIndex.  The values are stored in a single block in the PDOM record for
the QObject.

The API returns the value for a given key if it is found in the receiver
QObject or any of its base classes.  The API returns the first such
value that is found.

This patch also adds a test case for this functionality.

Change-Id: Ie3f821a0c5f6f1347a0c0c6dafa184510ae26c29
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>
Reviewed-on: https://git.eclipse.org/r/19154
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>
Tested-by: Doug Schaefer <dschaefer@qnx.com>