Bug 422841: Add Q_PROPERTY to the QtIndex
This extends IQObject so that it can return the list of Q_PROPERTY
expansions.  Each Q_PROPERTY is represented by the new IQProperty.  The
attributes of the property are stored in IQProperty.Attribute.

Where applicable, the attributes insert a reference from the associated
C++ binding.  This means that the Q_PROPERTY expansion will be included
in the list of references for the C++ binding.

This also simplifies the process for adding new PDOMBindings to the Qt
linkage.  New instances are stored in an implementation of IQtASTName
and able to directly create and return a new QtPDOMBinding.  This avoids
creating three parallel inheritance hierachies (compared to the previous
Qt PDOM implementations).

The patch includes test cases to check handling of Q_PROPERTY
expansions.

The patch also adjusts the PDOMNameTests test case to produce more
stable results.  There are three indexer triggers (two new files and an
explicit reindex), but only one call to waitForIndexer.  #waitForIndexer
sometimes returns early (if the indexer has not yet queued the work), so
the extra calls give it more chance to work properly.

Change-Id: I7d256d1a938d24a9eb726c472fb150a02f26eb32
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>