Update patch set 1
Patch Set 1:

The example I can think of is adding editor cursor point to the list of Scopes in the C/C++ Search dialog. In fact, I thought this was how the C/C++ Search dialog already worked.

It's pretty easy to see how you'd use this outside of the context of Qt. Just because we really need it for Qt doesn't make it specific to Qt.

Patch-set: 1