Update patch set 2
Patch Set 2: Code-Review+1

Is there any code checked in that implements IQMakeProjectInfoListener?  The part that I wanted to check wa to be sure the listener will not have problems if it is notified after it has been removed from the list.

I think that this could happen even before this patch, but it is something that I noticed now.  It could happen because there is no synchronization (and I don't think there should be) between the notify method and add/removeListener.  It just means that the implementers of IQMakeProjectInfoListener#qmakeInfoChanged need to handle being notified after they have removed themselves.

Secondly, I'm starting to rethink the wisdom of not passing event data to the notification method.  By not passing the IQMakeInfo all listeners will need to deal with duplicate notification where nothing has changed.  E.g., this sequence:

    Change-1
    Change-2
    Notify-for-change-1
    Notify-for-change-2

By the time listeners check the IQMakeInfo at notify-1, it will already been updated as the result of change-2.  They will then check it again (for notify-2) and get the same content.

Some listeners might care if change-2 undoes whatever change-1 did, but in this API there isn't a way for them to learn about that change.

I'm not certain that this matters, I'm just asking you if it does matter.

Finally, if you do change the notify method to pass an IQMakeInfo, then I think it should also be called as part of addListener:

public void addListener(IQMakeProjectInfoListener listener) {
    listeners.add(listener);
    listener.qmakeInfoChanged( copy_of_curr_info );
}

I don't see any problems with this particular change (which is why I +1 here).  I'm just asking questions that could be addressed in a followup change.

Patch-set: 2
Reviewer: Gerrit User 1042 <1042@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1